requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.9
streamlit>=1.35.0
streamlit-extras>=0.4.0
pydantic>=2.0.0
langchain-core>=0.2.0
langsmith>=0.1.77
python-dotenv>=1.0.0
